Fix issues for fuzzing
 - add cfg recipe to Makefile
 - add afl recipe which translates project to fuzzying ready binary (AFL fuzzer)
 - reduce debug output
 - unification of debug messages
 - unification of verbose messages
 - fix value contains instead integer a long integer

+++ src/ids.hpp
@@ -36,7 +36,7 @@ namespace st2se {
         val_format_t value_format;
         val_type_t value_type;
         std::string key;
-        std::variant<int, std::string> value;
+        std::variant<long, std::string> value;
         std::vector<_argument_t> next;
     };
